FinSight-AI/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
ddgs
ollama
langchain-chroma
langchain-ollama
yfinance
requests
langchain-text-splitters
pypdf